Fixes for crashes
These missing files which had to be under "/system/vendor/etc" folder leads to crashes on app which use sound. By placing them to where they have to be, crashes on music player when trying to play any music and video recording are fixed. Although there has been no sound yet, these changes are required in order to prevent crashes.

Also, there may be another way in order to fix crashes instead of doing these changes. It is to edit the paths to "/system/etc/" as the files are already placed there.

I guess the lines for "/system/etc" can be removed as well but I am not sure if there are any system part that use this directory. It's better to let them stay under "/system/etc" as well.