g4-common: selinux: wrild: complete re-done
after checking all relevant policies for wrild the most
of the stuff added in the past were bc of comm="ps" which
is not needed in most cases.

as we disabled neverallow we cannot use wildcards do deny those
silently. So dontaudit's have been added which likely needs to
be extended in the future.