g4-common: audio: fix for Q by switching to hwbinder
The root cause in Q which prevents audio from working was:

> msm_pcm_routing_get_app_type_idx: App type not available, fallback to default

So after investigating why the app type is not available I added debug code to the
calling functions and found that it will gets 48000 as the app_type.
When it compares that with the one from audio_output_policy.conf (which is 69937) it
fails as they do not match. so wth? After switching to hwbinder the app_type gets added
correctly by the kernel and audio works.

debugged function:
kernel/lge/msm8992/sound/soc/msm/qdsp6v2/msm-pcm-routing-v2.c
--> msm_routing_put_app_type_cfg_control
(the app_types get added here to the struct needed by the comparison)

Signed-off-by: steadfasterX <steadfasterX@gmail.com>