g4-common: sysctl: decrease extra_free_kbytes
default is 40 MB are hold by the kernel (per core),
decreasing that by 10 MB to not waste to much of RAM

In general this has been introduced to speed up real-time tasks.

This memory is kept free from use by normal processes. The special
processes that can run include interrupt routines,which must be run now
(as it's an interrupt), and finish before any other processes can run
(atomic). This can include things like swapping out memory to disk when
 memory is full.

If the memory is filled an interrupt (memory management) process runs
to swap some memory into disk so it can free some memory for use by
normal processes. But if vm.min_free_kbytes is too small for it to run,
then it locks up the system. This is because this interrupt process must
run first to free memory so others can run, but then it's stuck because it
doesn't have enough reserved memory vm.min_free_kbytes to do its
task resulting in a deadlock.

Change-Id: I1c88950d2124ee1a72c0c1b5031a4ee6ebc0f060