g4-common: wrild: fix sim-detection delay
due to the RIL fix we have implemented now that aggressive initial restart
of RILD isn't needed anymore.

this commit also implements a delay after boot (4min) before WRILD will begin
trying to fix RILD. usually that time should be more then enough to get RILD
boot properly but if not WRILD is here for the rescue.

WRILD is also still in the game to avoid the high-cpu-usage bug which can
(very!) rarely happen (might be not the case anymore.. but does not hurt to
check for it)

Signed-off-by: steadfasterX <steadfasterX@gmail.com>
Change-Id: I6b22e81b6ec035b6871de8be099e827f01e78300