Revert "g4-common: build missing Bluetooth A2DP"
* A2DP is no longer supported in A13

[  1% 1754/134015] build out/NOT_AVAILABLE_FOR_PLATFORM/target/product/h815/system/lib64/hw/audio.a2dp.default.so
FAILED: out/NOT_AVAILABLE_FOR_PLATFORM/target/product/h815/system/lib64/hw/audio.a2dp.default.so
/bin/bash -c "(echo -e \"\\033[1maudio.a2dp.default: \\033[31merror:\\033[0m\\033[1m\" 'Module is requested to be installed but is not available for platform because it does not have \"//apex_available:platform\" or it depends on other modules that are not available for platform.'  \"\\033[0m\" >&2 ) && (exit 1 )"
audio.a2dp.default: error: Module is requested to be installed but is not available for platform because it does not have "//apex_available:platform" or it depends on other modules that are not available for platform.
[  1% 1755/134015] build out/NOT_AVAILABLE_FOR_PLATFORM/target/product/h815/system/lib/hw/audio.a2dp.default.so
FAILED: out/NOT_AVAILABLE_FOR_PLATFORM/target/product/h815/system/lib/hw/audio.a2dp.default.so
/bin/bash -c "(echo -e \"\\033[1maudio.a2dp.default_32: \\033[31merror:\\033[0m\\033[1m\" 'Module is requested to be installed but is not available for platform because it does not have \"//apex_available:platform\" or it depends on other modules that are not available for platform.'  \"\\033[0m\" >&2 ) && (exit 1 )"
audio.a2dp.default_32: error: Module is requested to be installed but is not available for platform because it does not have "//apex_available:platform" or it depends on other modules that are not available for platform.
14:06:12 ninja failed with: exit status 1

This reverts commit 6f72e0132736d16c0816413ed0d24d807e97442f.